Move RateLimiting tests to separate file
